Add drag-and-drop functionality to CRUD component
Added new states to handle the index of the row that's being dragged and where it is to be dropped in the CRUD.js file. Removed the selectRow function which was previously used for row selection on click, and instead used onMouseDown event for the same on the rows. The drag-and-drop implementation was done using handleDragStart, handleDragOver, and handleDrop functions, and also, the changes were stored in local storage to maintain the changes.

In addition, two other XML files were created: 'Project_Default.xml' for setting up Eslint inspection profile with warning level and 'vcs.xml' for maintaining directory mapping for Git in the project.